+++ LDSContent/NavSection.swift
@@ -24,13 +24,13 @@ import Foundation
 
 public struct NavSection: Equatable {
     
-    public var id: Int
-    public var navCollectionID: Int
+    public var id: Int64
+    public var navCollectionID: Int64
     public var position: Int
     public var indentLevel: Int
     public var title: String?
     
-    public init(id: Int, navCollectionID: Int, position: Int, indentLevel: Int, title: String?) {
+    public init(id: Int64, navCollectionID: Int64, position: Int, indentLevel: Int, title: String?) {
         self.id = id
         self.navCollectionID = navCollectionID
         self.position = position
