+++ LDSContent/Source.swift
@@ -24,11 +24,11 @@ import Foundation
 
 public struct Source: Equatable {
     
-    public var id: Int
+    public var id: Int64
     public var name: String
     public var type: SourceType
     
-    public init(id: Int, name: String, type: SourceType) {
+    public init(id: Int64, name: String, type: SourceType) {
         self.id = id
         self.name = name
         self.type = type
